--- src/main/java/nextstep/auth/AuthenticationException.java
@@ -1,10 +1,0 @@
-package nextstep.auth;
-
-public class AuthenticationException extends RuntimeException {
-    public AuthenticationException() {
-    }
-
-    public AuthenticationException(String message) {
-        super(message);
-    }
-}
