--- src/main/java/nextstep/auth/application/TokenProvider.java
@@ -1,11 +1,0 @@
-package nextstep.auth.application;
-
-public interface TokenProvider {
-    boolean isSupport(TokenType tokenType);
-
-    String createToken(String principal);
-
-    String getPrincipal(String token);
-
-    boolean validateToken(String token);
-}
