--- src/main/java/nextstep/auth/client/ExternalTokenFetcher.java
@@ -1,10 +1,0 @@
-package nextstep.auth.client;
-
-import nextstep.auth.client.dto.ProfileResponse;
-
-public interface ExternalTokenFetcher {
-
-    String requestToken(String code);
-
-    ProfileResponse findUser(String accessToken);
-}
