--- src/main/java/nextstep/auth/client/github/dto/GithubAccessTokenRequest.java
@@ -1,27 +1,0 @@
-package nextstep.auth.client.github.dto;
-
-public class GithubAccessTokenRequest {
-    private String code;
-    private String clientId;
-    private String clientSecret;
-
-    public GithubAccessTokenRequest(String code,
-                                    String clientId,
-                                    String clientSecret) {
-        this.code = code;
-        this.clientId = clientId;
-        this.clientSecret = clientSecret;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public String getClientId() {
-        return clientId;
-    }
-
-    public String getClientSecret() {
-        return clientSecret;
-    }
-}
