--- src/main/java/nextstep/auth/domain/LoginMember.java
@@ -1,13 +1,0 @@
-package nextstep.auth.domain;
-
-public class LoginMember {
-    private String email;
-
-    public LoginMember(String email) {
-        this.email = email;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-}
