+++ src/main/java/nextstep/subway/auth/AuthenticationException.java
@@ -1,0 +1,10 @@
+package nextstep.subway.auth;
+
+public class AuthenticationException extends RuntimeException {
+    public AuthenticationException() {
+    }
+
+    public AuthenticationException(String message) {
+        super(message);
+    }
+}
