+++ src/main/java/nextstep/subway/auth/application/TokenProvider.java
@@ -1,0 +1,11 @@
+package nextstep.subway.auth.application;
+
+public interface TokenProvider {
+    boolean isSupport(TokenType tokenType);
+
+    String createToken(String principal);
+
+    String getPrincipal(String token);
+
+    boolean validateToken(String token);
+}
