+++ src/main/java/nextstep/subway/auth/client/ExternalTokenFetcher.java
@@ -1,0 +1,10 @@
+package nextstep.subway.auth.client;
+
+import nextstep.subway.auth.client.dto.ProfileResponse;
+
+public interface ExternalTokenFetcher {
+
+    String requestToken(String code);
+
+    ProfileResponse findUser(String accessToken);
+}
