+++ src/main/java/nextstep/subway/auth/client/github/dto/GithubAccessTokenRequest.java
@@ -1,0 +1,27 @@
+package nextstep.subway.auth.client.github.dto;
+
+public class GithubAccessTokenRequest {
+    private String code;
+    private String clientId;
+    private String clientSecret;
+
+    public GithubAccessTokenRequest(String code,
+                                    String clientId,
+                                    String clientSecret) {
+        this.code = code;
+        this.clientId = clientId;
+        this.clientSecret = clientSecret;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public String getClientId() {
+        return clientId;
+    }
+
+    public String getClientSecret() {
+        return clientSecret;
+    }
+}
