+++ src/main/java/nextstep/subway/auth/domain/LoginMember.java
@@ -1,0 +1,13 @@
+package nextstep.subway.auth.domain;
+
+public class LoginMember {
+    private String email;
+
+    public LoginMember(String email) {
+        this.email = email;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+}
