+++ src/main/java/nextstep/subway/favorite/ui/FavoriteController.java
@@ -1,11 +1,11 @@
 package nextstep.subway.favorite.ui;
 
 
-import nextstep.auth.ui.AuthenticationPrincipal;
+import nextstep.subway.auth.ui.AuthenticationPrincipal;
 import nextstep.subway.favorite.application.FavoriteService;
 import nextstep.subway.favorite.application.dto.FavoriteRequest;
 import nextstep.subway.favorite.application.dto.FavoriteResponse;
-import nextstep.auth.domain.LoginMember;
+import nextstep.subway.auth.domain.LoginMember;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
