+++ src/main/java/nextstep/subway/handler/ApiExceptionHandler.java
@@ -1,6 +1,6 @@
 package nextstep.subway.handler;
 
-import nextstep.auth.AuthenticationException;
+import nextstep.subway.auth.AuthenticationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
