+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -1,10 +1,10 @@
 package nextstep.subway.member.ui;
 
-import nextstep.auth.ui.AuthenticationPrincipal;
+import nextstep.subway.auth.ui.AuthenticationPrincipal;
 import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.application.dto.MemberRequest;
 import nextstep.subway.member.application.dto.MemberResponse;
-import nextstep.auth.domain.LoginMember;
+import nextstep.subway.auth.domain.LoginMember;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
