+++ src/test/java/nextstep/subway/member/acceptance/TokenAcceptanceTest.java
@@ -3,7 +3,7 @@ package nextstep.subway.member.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.auth.application.TokenType;
+import nextstep.subway.auth.application.TokenType;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import nextstep.subway.testhelper.AcceptanceTest;
